+++ src/cli/export/export.go
@@ -6,8 +6,6 @@ import (
 
 var exportCommands = []func(cmd *cobra.Command) *cobra.Command{
 	addOneDriveCommands,
-	// addExchangeCommands,
-	// addSharePointCommands,
 }
 
 // AddCommands attaches all `corso export * *` commands to the parent.