+++ src/cli/flags/export.go
@@ -11,5 +11,5 @@ var ArchiveFV bool
 // AddExportConfigFlags adds the restore config flag set.
 func AddExportConfigFlags(cmd *cobra.Command) {
 	fs := cmd.Flags()
-	fs.BoolVar(&ArchiveFV, ArchiveFN, false, "Archive contents")
+	fs.BoolVar(&ArchiveFV, ArchiveFN, false, "Export data as an archive instead of individual files")
 }