+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SharePoint document libraries deleted after the last backup can now be restored.
 
 ### Added
-- Added option to export data from OneDrive backups
+- Added option to export data from OneDrive backups as individual files or as a single zip file.
 
 ## [v0.11.1] (beta) - 2023-07-20